requirements.txt
autoawq==0.2.7.post3
bitsandbytes
python-multipart
langchain
langchain-core
langchain-chroma
langchain-community
uvicorn
fastapi
vllm
pypdf
torch==2.5.1
--extra-index-url https://download.pytorch.org/whl/cu121
python-dotenv
sentence-transformers==3.3.1
streamlit==1.41.1